modification

+++ project/src/main/java/com/controller/controller/UserController.java
@@ -22,15 +22,13 @@ public class UserController {
     }
 
     // build get user by id REST API
-    // http://localhost:8080/api/users/1
     @GetMapping("{id}")
-    public ResponseEntity<User> getUserById(@PathVariable("id") Long userId){
+    public ResponseEntity<User> getUserById(@PathVariable("id") int userId){
         User user = userService.getUserById(userId);
         return new ResponseEntity<>(user, HttpStatus.OK);
     }
 
     // Build Get All Users REST API
-    // http://localhost:8080/api/users
     @GetMapping
     public ResponseEntity<List<User>> getAllUsers(){
         List<User> users = userService.getAllUsers();
@@ -39,8 +37,7 @@ public class UserController {
 
     // Build Update User REST API
     @PutMapping("{id}")
-    // http://localhost:8080/api/users/1
-    public ResponseEntity<User> updateUser(@PathVariable("id") Long userId,
+    public ResponseEntity<User> updateUser(@PathVariable("id") int userId,
                                            @RequestBody User user){
         user.setId(userId);
         User updatedUser = userService.updateUser(user);
@@ -49,7 +46,7 @@ public class UserController {
 
     // Build Delete User REST API
     @DeleteMapping("{id}")
-    public ResponseEntity<String> deleteUser(@PathVariable("id") Long userId){
+    public ResponseEntity<String> deleteUser(@PathVariable("id") int userId){
         userService.deleteUser(userId);
         return new ResponseEntity<>("User successfully deleted!", HttpStatus.OK);
     }
